+++ backend/superprice/src/test/java/com/example/superprice/controllers/SuperpriceControllerTest.java
@@ -42,12 +42,12 @@ public class SuperpriceControllerTest {
     void should_returnProduct_When_relatedProducts() {
         Product p1 = new Product((long) 4, "Coke",
                 "A totally healthy beverage that is very tasty.", "Woolworths",
-                "/cokeBottle.png", (long) 69, (long) 6);
+                "/cokeBottle.png", 69, 6);
 
         when(this.service.findByKeyword("Coke")).thenReturn(
                 List.of(new Product((long) 4, "Coke",
                         "A totally healthy beverage that is very tasty.", "Woolworths",
-                        "/cokeBottle.png", (long) 69, (long) 6)));
+                        "/cokeBottle.png", 69, 6)));
 
         Collection<Product> p = this.controller.searchForProduct("Coke");
         assertNotNull(p);
@@ -56,26 +56,26 @@ public class SuperpriceControllerTest {
 
     // Get all tests
     @Test
-    void should_returnCollection_When_getProducts () {
+    void should_returnCollection_When_getProducts() {
         Product p1 = new Product((long) 1, "T-Shirt",
                 "This is a pretty cool shirt.", "Coles",
-                "/tshirt.png", (long) 19.99, (long) 20);
+                "/tshirt.png", 19.99, 20);
 
         Product p2 = new Product((long) 2, "Coke",
                 "A totally healthy beverage that is very tasty.", "Woolworths",
-                "/cokeBottle.png", (long) 3.99, (long) 18);
+                "/cokeBottle.png", 3.99, 18);
 
         Product p3 = new Product((long) 3, "Molten Basketball",
                 "A nice basketball that is not overpriced at all.", "Woolworths",
-                "/basketball.png", (long) 69, (long) 6);
+                "/basketball.png", 69, 6);
 
         Product p4 = new Product((long) 4, "Samsung",
                 "A new smart watch which is totally necessary.", "Coles",
-                "/watch.webp", (long) 200, (long) 3);
+                "/watch.webp", 200, 3);
 
         Product p5 = new Product((long) 5, "Coke",
                 "A totally healthy beverage that is very tasty.", "Woolworths",
-                "/cokeBottle.png", (long) 3.99, (long) 18);
+                "/cokeBottle.png", 3.99, 18);
 
         when(this.service.getProducts()).thenReturn(
                 List.of(p1, p2, p3, p4, p5));
@@ -105,7 +105,7 @@ public class SuperpriceControllerTest {
         when(this.service.getCartProducts(1)).thenReturn(
                 List.of(new Product((long) 4, "Coke",
                         "A totally healthy beverage that is very tasty.", "Woolworths",
-                        "/cokeBottle.png", (long) 69, (long) 6)));
+                        "/cokeBottle.png", 69, 6)));
         Collection<Product> p = this.controller.getCartProducts(1);
         assertEquals(1, p.size());
     }
